Adapt examples to new macro names
